MS Scraper V1.0
This scrapes a website for Company names and addresses including
telephone number and stores it in a csv file. Runs using Tor as the
anonymous proxy and Privoxy to run it through http. Uses Python Scraper
to pull and extract the data.

Included is a full installer script to install all dependencies
(assuming on Ubuntu 12.04) and adjusts the config, then restarts the
services.

Known bug: Tor socket configured to the same password each time when
using the install script. To be fixed in next update.